project/requirements.txt
accelerate==0.32.1
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.3
beanie==1.26.0
beautifulsoup4==4.12.3
billiard==4.2.0
black==24.4.2
blinker==1.8.2
build==1.2.1
cachetools==5.3.3
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.5
chromadb==0.5.4
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coloredlogs==15.0.1
dataclasses-json==0.6.7
datasets==2.20.0
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
dill==0.3.8
dnspython==2.6.1
ecdsa==0.19.0
einops==0.8.0
email_validator==2.2.0
emoji==2.12.1
exceptiongroup==1.2.1
executing==2.0.1
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.15.4
filetype==1.2.0
Flask==3.0.3
Flask-Cors==4.0.1
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.5.0
google-auth==2.32.0
googleapis-common-protos==1.63.2
greenlet==3.0.3
grpcio==1.64.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
idna==3.7
importlib_metadata==7.1.0
importlib_resources==6.4.0
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==3.0.0
kombu==5.3.7
kubernetes==30.1.0
langchain==0.2.7
langchain-chroma==0.1.2
langchain-community==0.2.7
langchain-core==0.2.18
langchain-huggingface==0.0.3
langchain-openai==0.1.16
langchain-text-splitters==0.2.2
langchainhub==0.1.20
langdetect==1.0.9
langsmith==0.1.85
lazy-model==0.2.0
llvmlite==0.43.0
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
matplotlib-inline==0.1.7
mdurl==0.1.2
mmh3==4.1.0
motor==3.4.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.5.82
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
ollama==0.2.1
onnxruntime==1.18.1
openai==1.35.14
openai-whisper==20231117
opentelemetry-api==1.25.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-asgi==0.46b0
opentelemetry-instrumentation-fastapi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
ordered-set==4.1.0
orjson==3.10.6
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parso==0.8.4
passlib==1.7.4
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
platformdirs==4.2.2
posthog==3.5.0
prompt_toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
pure-eval==0.2.2
pyarrow==17.0.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-settings==2.3.4
pydantic_core==2.20.1
Pygments==2.18.0
pymongo==4.7.3
pypdf==4.2.0
PyPika==0.48.9
pyproject_hooks==1.1.0
PySoundFile==0.9.0.post1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.0.1
python-iso639==2024.4.27
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.9.4
redis==5.0.7
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.7.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
sentence-transformers==3.0.1
sentencepiece==0.2.0
shellingham==1.5.4
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.5
SQLAlchemy==2.0.31
stack-data==0.6.3
starlette==0.37.2
sympy==1.13.0
tabulate==0.9.0
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
toml==0.10.2
tomli==2.0.1
torch==2.3.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.42.4
triton==2.3.1
typer==0.12.3
types-requests==2.32.0.20240622
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
ujson==5.10.0
unstructured==0.14.10
unstructured-client==0.24.1
urllib3==2.2.2
uvicorn==0.30.1
uvloop==0.19.0
vine==5.1.0
watchfiles==0.22.0
wcwidth==0.2.13
websocket-client==1.8.0
Werkzeug==3.0.3
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.19.2
